lots of previously uncommitted changes...

+++ filtFictrac_all.m
@@ -25,9 +25,9 @@ function filtFictrac_all()
 
     % some constants
     fictracParams.dsf = 20; % downsample to 1000 Hz;
-    fictracParams.filtParams.padLen = int32(200);
-    fictracParams.filtParams.sigmaPos = int32(100); % 100 ms
-    fictracParams.filtParams.sigmaVel = int32(50); % 50 ms
+    fictracParams.filtParams.padLen = 200;
+    fictracParams.filtParams.sigmaPos = 100; % 100 ms
+    fictracParams.filtParams.sigmaVel = 50; % 50 ms
     
     BALL_DIAM = 6.46;
     circum = BALL_DIAM * pi; % circumference of ball, in mm
